+++ classes/Gems/Tracker/Engine/TrackEngineAbstract.php
@@ -115,7 +115,7 @@ abstract class Gems_Tracker_Engine_TrackEngineAbstract extends \MUtil_Translate_
                 ->where('gro_id_track = ?', $this->_trackId)
                 ->order('gro_id_order');
 
-            // \MUtil_Echo::r((string) $roundSelect);
+            // \MUtil_Echo::track((string) $roundSelect, $this->_trackId);
 
             $this->_rounds  = array();
             foreach ($roundSelect->query()->fetchAll() as $round) {